requirements.txt
annotated-types==0.7.0
anyio==4.4.0
assemblyai==0.26.0
asttokens==2.4.1
certifi==2024.6.2
charset-normalizer==2.0.12
coloredlogs==15.0.1
decorator==5.1.1
distro==1.9.0
EfficientWord-Net==1.0.2
elevenlabs==1.2.2
exceptiongroup==1.2.1
executing==2.0.1
flatbuffers==24.3.25
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
humanfriendly==10.0
idna==3.7
ipython==8.18.1
jedi==0.19.1
matplotlib-inline==0.1.7
mpmath==1.3.0
numpy==1.22.0
onnxruntime==1.18.0
openai==1.30.5
packaging==24.0
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.45
protobuf==5.27.0
ptyprocess==0.7.0
pure-eval==0.2.2
PyAudio==0.2.11
pydantic==2.7.2
pydantic_core==2.18.3
Pygments==2.18.0
python-dotenv==1.0.1
requests==2.26.0
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
sympy==1.12.1
tflite==2.10.0
tflite-runtime==2.5.0.post1
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.12.1
urllib3==1.26.18
wcwidth==0.2.13
websockets==12.0
